Extract coursier scala installation setup & cleanup for tests to a helper trait for reuse
